requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
autopep8==1.7.0
click==8.1.3
fastapi==0.85.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
pycodestyle==2.9.1
pydantic==1.10.2
pypika-tortoise==0.1.6
pytz==2022.4
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.3
